chore(mockingbird-ts): move generator into dedicated folder
